requirements.txt
asgiref==3.11.1
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.12.1
python-decouple==3.8
pytz==2026.1.post1
sqlparse==0.5.5
tzdata==2025.3
gunicorn==21.2.0
whitenoise==6.6.0
uvicorn==0.24.0
channels==4.0.0
channels-redis==4.1.0
